Fix location of data directory
This was unintentional but the data directory was actually being created _inside_ the source directory. The only reason the test didn't fail is because I was building up the filessytem in `assertFiles` with `Object.assign`, which does not use a deep merge. That means it would have been possible for `assertFiles` to assert against multiple files because the last one was overwriting the previous one.