Refactor building to be asynchronous
I would have loved for it to be natively synchronous, but while postcss does support sync behavior, many of its plugins do not, and if one postcss plugin is async, then your out of luck, you have to use async.

I can get back the sync behavior by using async/await, but that appears to cause some problems with the tests. The tests are now no longer synchronous and so will have to be run serially. There's also an issue with the `afterEach` hook not running at the run time—it seems to run before the end of test, inexplicably, and this might be a known issue. To work around it I'm just manually calling the cleanup function at the end of each test.