Refactor file building into the file itself
This lets me move the last decision about file type into the file itself. Eventually this will allow me to create different file classes that wrap up the behavior specific to each class, which will reduce the number of switches on file type to one (which class to instantiate for each file).