Switch from babel-register to a manual override
Unfortunately babel-register does not provide enough flexibility. It’s not possible to do different transforms for different file types, or do transformations _before_ babel has a chance to do transformations. Something which will be necessary for the next feature (MDX support).

pirates is the method of overriding `require`that babel-register uses, so we can use that directly to get the necessary flexibility.