Manually resolve module paths for Babel plugins and presets
There appears to be a bug with babel when using `npm link` (a pattern for working on packages locally) where the plugins and presets that you pass in the config are resolved relative to the wrong directory. It’s unclear if/when this will be fixed, but in the meantime a workaround is to manually resolve them.