Add a TextFile class that converts files to a string
This fixes a goof where _all_ files were being converted to a string while being built. This is fine and necessary for text files that need some form of transformation done, like JSX, JS, CSS, etc. but this unfortunately caused a bug with non-text files like images, fonts, etc. because they were being converted to a string and then written back to disk, essentially corrupting them.

The change to how the static (class) property `fileTypes` is being defined on `File` is because of a subtle bug where the property was being defined on a sub-class of `File` rather than `File` itself. This didn’t manifest itself before because coincidentally `registerFileType` was always first called on `File` before it was called on a sub-class like `HTMLFile`. But adding `TextFile` inherently called `registerFileType` on `TextFile` first, creating `fileTypes` on `TextFile`, thus making `File.fileTypes` `undefined` and causing `File.instantiateByType` to fail.

Defining `fileTypes` during load fixes the problem and in the future when the “class fields” proposal moves from Stage 3 to Stage 4 it can be defined in the class body itself.