Clear require cache before building
This is a subtle bug where because the way JSX/MDX files get built is by requiring the files, if the file changes, the next time a build happens it will pull the stale version from the cache, not the updated version.

I ran into this initially when running tests. It cropped up again when running the server. Because the server watches files and rebuilds when a file is changed, without ever exiting, every time it rebuilds we need to clear the require cache first to ensure we don't keep getting the stale versions of the files when they’re required.