Switch from cssnext to postcss-preset-env
cssnext is deprecated in favor of postcss-preset-env.

cssnext took an approach of implementing almost any proposed CSS feature regardless of its stability or likelihood of eventually becoming part of the spec. postcss-preset-env takes a different approach, allowing you to specify the “stage” at which you want to include experimental features.

postcss-preset-env recommends Stage 2, which I agree with, although I think also enabling `custom-media-queries` from Stage 1 makes sense. It seems likely to make it to Stage 2 (and onward) and a lot of people (myself included) use it.