Fix a gnarly recursive bug
This `filter` function tries to find the files that import the file being built. There’s a try/catch because if the file has a parsing error, then it won't be possible to parse the file in order to find which files it imports. In the `catch` block we were incorrectly returning an array, rather than `true` or `false`, which is what `filter` expects. I think I probably got confused and was thinking about the ultimate return value of the `filter` function, which is an array.

An array is a truthy value, which means every single file would have been returned from the function, including the file itself, which would cause an infinite loop.

The correct thing to do is return `false`.