⬆️ babel@7
This is a rather large upgrade. Most of it is changing package names to use the “@babel” scope, but there are some small behavior changes as well. I added a few tests to the JSX and MDX import parsers to make it easier to see errors that were being swallowed by the integration tests.