Watch data directory for changes and reload browser
This requires slightly changing the logged message when reloading the browser.

Initially I was logging which file had changed as the file that was being “reloaded”. Which sort of worked when the files that were being watched were source files, but watching a data file means the page should be reloaded, but there’s no way to know which page is currently being viewed. So it makes more sense to just not specify what is being reloaded.