Fix bug with server not loading non-root index files
The Browsersync server doesn’t work how I thought it did. When making a request to `some-page` I thought it could be configured to read the file from `some-page/index.html`, but it can’t.

I added a middleware to handle this and also added some tests to ensure it works.