Change how sites are built for clean URLs
Originally, Charge attempted to support clean URLs by building HTML files in a specific way. For example, a file named `/projects.html` would be built as `/projects/index.html`. In retrospect, this was a big mistake.

The problem is every host handles clean URLs and directory indexes differently. Some will read the structure that Charge outputs and do what people probably expect but some won’t. Fundamentally, clean URLs is a choice about how to map a URL to a static file, and that’s really a concern of the host, not of the tool that builds the static site.

So I’m removing the code that builds HTML files with directory indexes and adding support to the dev server to load HTML files both with and without extensions.