Simplify filesystem comparisons
Originally the comparisons were done as nested objects (a tree structure). This required globbing all the files in the target directory and then expanding them into the nested structure so it could be compared against the expected files in the same structure.

In retrospect, it’s simpler to glob the files, leave them in the “flat” structure, and then flatten the expected files.

This might also make failed test output a bit easier to read.