Switch from Signale to a custom logger
Signale is great but it doesn’t have the ability to align logger scopes/groups, which I think makes the output look weird. I considered adding the feature to Signale, but the way it is designed makes it pretty difficult. I was only using Signale in a really superficial way, so it wound up just being much easier to implement a simple logger myself.

Fixes #1