Fix tests on Windows
The problem is that the snapshots of the filesystem are generated on *nix (with forward slashes) so when the tests are run on Windows (which uses back slashes) the tests fail.

What’s not clear is why this worked before and then suddenly started failing.  The commits around when it started failing don’t seem related in any way 🤷🏼‍♂️